backend/requirements.txt
# Flask Framework
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0

# Database
PyMySQL==1.1.0
cryptography==41.0.7

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Security
Werkzeug==3.0.1

# Production Server
gunicorn==21.2.0

# Development Tools (optional)
pytest==7.4.3
pytest-flask==1.3.0